feature: Score 따로 빼는 저략
