+++ lms-core/src/main/java/com/savelms/core/attendance/domain/CheckIn.java
@@ -1,0 +1,5 @@
+package com.savelms.core.attendance.domain;
+
+public enum CheckIn {
+    ABSENT, ATTEND, VACATION
+}
