+++ lms-core/src/main/java/com/savelms/core/attendance/domain/CheckOut.java
@@ -1,0 +1,5 @@
+package com.savelms.core.attendance.domain;
+
+public enum CheckOut {
+    ABSENT, ATTEND, VACATION
+}
