+++ lms-core/src/main/java/com/savelms/core/calendar/domain/entity/Calendar.java
@@ -1,33 +1,32 @@
-package com.savelms.core;
+package com.savelms.core.calendar.domain.entity;
 
+import com.savelms.core.BaseEntity;
+import com.savelms.core.calendar.DayType;
+import com.savelms.core.attendance.domain.entity.Attendance;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
-
-//@SequenceGenerator(
-//    name = "TODO_GENERATOR",
-//    sequenceName = "MEMBER_SEQ",
-//    initialValue = 1,
-//    allocationSize = 50)
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
-@Table(name = "TODO")
+@Getter
+@Table(name = "CALENDAR")
 @Entity
-public class Todo extends BaseEntity{
-
+@Builder
+public class Calendar extends BaseEntity {
 
 
     //********************************* static final 상수 필드 *********************************/
@@ -40,24 +39,23 @@ public class Todo extends BaseEntity{
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="TODO_ID")
+    @Column(name="CALENDAR_ID")
     private Long id;
 
     /********************************* PK가 아닌 필드 *********************************/
+    LocalDate date;
 
-    private boolean complete;
-    private String title;
-
-    /********************************* 연관관계 매핑 *********************************/
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="ATTENDANCE_ID")
-    private Attendance attendance;
-
+    @Enumerated(value = EnumType.STRING)
+    DayType dayType;
 
 
     /********************************* 비영속 필드 *********************************/
 
+
+    /********************************* 연관관계 매핑 *********************************/
+
+    @OneToMany(mappedBy = "calendar")
+    private List<Attendance> attendances = new ArrayList<>();
 
 
     /********************************* 비니지스 로직 *********************************/
