+++ lms-core/src/main/java/com/savelms/core/user/domain/entity/TeamEnum.java
@@ -1,0 +1,5 @@
+package com.savelms.core.user.domain.entity;
+
+public enum TeamEnum {
+    RED, BLUE
+}
