+++ lms-core/src/main/java/com/savelms/core/user/domain/entity/UserRole.java
@@ -1,16 +1,16 @@
-package com.savelms.core;
+package com.savelms.core.user.domain.entity;
 
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
+import com.savelms.core.BaseEntity;
+import com.savelms.core.user.domain.entity.Role;
+import com.savelms.core.user.domain.entity.User;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,11 +20,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-@Table(name = "CALENDAR")
+@Table(name = "USER_ROLE")
 @Entity
 @Builder
-public class Calendar extends BaseEntity{
-
+public class UserRole extends BaseEntity {
 
     //********************************* static final 상수 필드 *********************************/
 
@@ -36,14 +35,10 @@ public class Calendar extends BaseEntity{
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="CALENDAR_ID")
+    @Column(name = "USER_ROLE_ID")
     private Long id;
 
     /********************************* PK가 아닌 필드 *********************************/
-    LocalDate date;
-
-    @Enumerated(value = EnumType.STRING)
-    DayType dayType;
 
 
     /********************************* 비영속 필드 *********************************/
@@ -51,12 +46,15 @@ public class Calendar extends BaseEntity{
 
     /********************************* 연관관계 매핑 *********************************/
 
-    @OneToMany(mappedBy = "calendar")
-    private List<Attendance> attendances = new ArrayList<>();
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="USER_ID")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="ROLE_ID")
+    private Role role;
 
     /********************************* 비니지스 로직 *********************************/
-
-
 
 }
